+++ src/main/java/com/crud/crudsimple/models/Pais.java
@@ -25,7 +25,7 @@ public class Pais {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "pai_id", unique = true, nullable = false, updatable = false)
-	private int idPais;
+	private Long idPais;
 
 	@Column(name = "pai_nome", nullable = false, length = 20)
 	@NotNull(groups = {CreatePais.class, UpdatePais.class})
